docs(server): PerTenantConfig enforces none of its five limits
All five TenantLimits fields are stored, resolvable through
PerTenantConfig::get, and read by no code in the request path. Verified by
enumerating every mention of the type outside its own module: an import, a
field, the tenant_config() accessor, a Debug line, a builder field and its
setter, and otherwise tests. Nothing calls .get() on it while serving a
request.

executor_timeout and event_queue_capacity are invisible to a name-based sweep
because they are also the names of live process-wide fields on RequestHandler
and RequestHandlerBuilder — messaging/mod.rs:487 reads self.executor_timeout,
the handler's own, and applies it to every tenant alike. Those fields are not
these fields, and the collision is what made the whole struct look wired.

What makes this the worst instance of the shape so far is not the count. The
module header advised operators to "set per-tenant max_concurrent_tasks so that
the sum across active tenants stays within the process-wide caps if
noisy-neighbor isolation matters" — operational advice to rely on a field
nothing reads, under a heading about fairness. An operator who followed it did
the work and came away believing they had isolation they did not have. A
missing feature is visible; a false one is not.

Enforcing all five is a feature with open design questions in a published API:
where a per-tenant semaphore lives and what exceeding it returns, how a
per-tenant rate limit agrees with RateLimitInterceptor rather than races it,
and a per-tenant store cap the tenant-aware stores cannot currently express
since they hand one per_tenant config to every partition. Enforcing *some* is
the trap and is the exact defect class being catalogued — executor_timeout is a
few lines, and shipping it alone would leave one live knob beside four inert
ones. So behaviour is unchanged and the documentation is now exactly true: the
module says it stores and resolves, each field names what would enforce it, and
the two builder setters no longer claim to enable per-tenant limits. Filed as
B22, scoped with its design questions named.

Data isolation is unaffected — it runs through the tenant-aware stores'
partitioning, not through any limit here. What is missing is fairness, not
separation.

tenant_limits_are_stored_and_resolved_but_never_applied pins the gap: a tenant
executor_timeout of 999s against a handler-wide 11s, asserting the handler
still runs at 11s. It dies under both a handler that drops its own timeout and
a get() that ignores overrides. When B22 lands it fails on purpose — the
failure is the reminder that this paragraph and five field notes became false
in the same change.

handler/mod.rs crossed the 500-line ratchet at 553 with that test, so its tests
move to handler/tests.rs, the seam rate_limit/shared_tests.rs already uses. 773
tests before and after.
